+++ changes/136.misc.md
@@ -1,0 +1,1 @@
+The Qilisim C++ code has been significantly refactored to improve compilation times and readability.
